+++ src/test/java/me/desair/tus/server/HttpProblemDetailsTest.java
@@ -54,6 +54,27 @@ public class HttpProblemDetailsTest {
                 + "\"provided-offset\":25000000}"));
   }
 
+  @Test
+  public void testOffsetMismatchProblemDetailsNullProvidedOffset() throws Exception {
+    HttpProblemDetails problem = HttpProblemDetails.forOffsetMismatch(12500000L, null);
+
+    assertThat(problem.getStatus(), is(409));
+    assertThat(problem.getExtraFields().get("expected-offset"), is(12500000L));
+    assertThat(problem.getExtraFields().containsKey("provided-offset"), is(false));
+
+    problem.writeTo(new TusServletResponse(response));
+
+    assertThat(response.getStatus(), is(409));
+    assertThat(
+        response.getContentAsString(),
+        is(
+            "{\"type\":\"https://iana.org/assignments/http-problem-types#mismatching-upload-offset\","
+                + "\"title\":\"Offset Mismatch\","
+                + "\"status\":409,"
+                + "\"detail\":\"The provided Upload-Offset does not match the server's current offset\","
+                + "\"expected-offset\":12500000}"));
+  }
+
   /**
    * Section 7.2 (Completed Upload) of draft-ietf-httpbis-resumable-upload-12 & RFC 7807: "This
    * section defines the 'https://iana.org/assignments/http-problem-types#completed-upload' problem