+++ src/test/java/me/desair/tus/server/rufh/handler/RufhUploadLimitHeaderRequestHandlerTest.java
@@ -166,4 +166,45 @@ public class RufhUploadLimitHeaderRequestHandlerTest {
     assertThat(uploadLimit, is(notNullValue()));
     assertThat(uploadLimit.contains("max-age=120"), is(true));
   }
+
+  @Test
+  public void testProcessWithLocationHeader() throws Exception {
+    when(uploadStorageService.getMaxUploadSize()).thenReturn(10000L);
+    when(uploadStorageService.getUploadInfo("/files/123", null)).thenReturn(null);
+
+    servletResponse.setHeader(HttpHeader.LOCATION, "/files/123");
+    TusServletResponse tusResponse = new TusServletResponse(servletResponse);
+
+    handler.process(HttpMethod.POST, null, tusResponse, uploadStorageService, null);
+
+    assertThat(tusResponse.getHeader(HttpHeader.UPLOAD_LIMIT), is("max-size=10000"));
+  }
+
+  @Test
+  public void testProcessGetUploadInfoThrowsException() throws Exception {
+    when(uploadStorageService.getMaxUploadSize()).thenReturn(10000L);
+    when(uploadStorageService.getUploadInfo(nullable(String.class), nullable(String.class)))
+        .thenThrow(new RuntimeException("Storage failure"));
+
+    TusServletResponse tusResponse = new TusServletResponse(servletResponse);
+
+    handler.process(
+        HttpMethod.POST,
+        new TusServletRequest(servletRequest),
+        tusResponse,
+        uploadStorageService,
+        null);
+
+    assertThat(tusResponse.getHeader(HttpHeader.UPLOAD_LIMIT), is("max-size=10000"));
+  }
+
+  @Test
+  public void testProcessWithBlankUriAndNullRequest() throws Exception {
+    when(uploadStorageService.getMaxUploadSize()).thenReturn(10000L);
+    TusServletResponse tusResponse = new TusServletResponse(servletResponse);
+
+    handler.process(HttpMethod.POST, null, tusResponse, uploadStorageService, null);
+
+    assertThat(tusResponse.getHeader(HttpHeader.UPLOAD_LIMIT), is("max-size=10000"));
+  }
 }